Feat addind validations

+++ src/genre/genre.service.ts
@@ -9,11 +9,20 @@ export class GenreService {
   }
 
   public list(): IGenre[] {
-    return this.genreRepository.findAll();
+    const genres = this.genreRepository.findAll()
+
+    if (!genres || genres.length === 0) {
+      throw new Error("No genres found in database.");
+    }
+    return genres;
   }
 
   public getById(id: number): IGenre | undefined {
-    return this.genreRepository.findById(id);
+    const genre = this.genreRepository.findById(id);
+    if (!genre) {
+      throw new Error(`Genre with ${id} not found in database.`);
+    }
+    return genre;
   }
 
   public create(name: string): IGenre {
@@ -21,10 +30,19 @@ export class GenreService {
   }
 
   public update(genre: IGenre): IGenre {
+    const genreUpdate = this.genreRepository.findById(genre.id);
+    if (!genreUpdate) {
+      throw new Error(`Genre not found in database. Unable to update.`)
+    }
     return this.genreRepository.update(genre);
   }
 
   public delete(id: number): void {
+    const genre = this.genreRepository.findById(id);
+
+    if (!genre) {
+      throw new Error(`Genre with ${id} not found in database. Unable to delete.`);
+    }
     this.genreRepository.delete(id);
   }
 }
